requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.7
django-livereload-server==0.3.2
django-webp==2.0.0
django-webp-converter==0.3.0
djangorestframework==3.12.4
httpagentparser==1.9.1
Jinja2==2.11.3
MarkupSafe==1.1.1
Pillow==8.2.0
pytz==2021.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tornado==6.1
watchdog==2.0.2
websocket-client==0.58.0
